requirements.txt
pillow

pyqt5
